src/requirements.txt
pendulum
apache-airflow
pandas
pymongo
'apache-airflow[postgres]'